wm: Add support for forcing compositing.
Things are currently glitchy if we exit the non-composited
mode by resizing a fullscreen window to instead just cover a
portion of the screen, since the X window often gets resized
before we get a chance to redraw and show the composited
scene.

This change adds an IncrementCompositingRequests() method to
WindowManager that can be used to force us to composite (and
that triggers a single composited draw if we're not already
compositing).  This can be called before we do a resize to
ensure that the ugliness that goes on in the uncomposited
world isn't visible onscreen.  At this point, we'll continue
compositing while we wait for the resized window to be
repainted by the client.

BUG=chromium-os:13963
TEST=manually tested it with the new layout code that i'm working on and checked that there's no ugly flicker when i go from maximized to non-maximized mode; also added some tests

Review URL: http://codereview.chromium.org/6843001

Change-Id: I05cab9fa1e0080d261a6484c2734f618cf70e6e7