wm: Disable Ctrl-Alt-T until first browser window is up.
There's a weird state the first time a new user logs in
where the _CHROME_LOGGED_IN property is set on the root
window but the camera wizard is still being shown.  This
change moves the Ctrl-Alt-T shortcut for opening a terminal
from WindowManager to LayoutManager, so we can disable it
until the user has finished taking their picture and a
browser window is being shown.

Change-Id: I4c1ddf3349cfcc5662306e1f3d5a607c6eff5304

BUG=chromium-os:12118
TEST=logged in with a new user and checked that ctrl-alt-t did nothing on the camera screen but worked afterwards

Review URL: http://codereview.chromium.org/6854010