wm: Synchronize all animations added in the same frame.
When scheduling a new animation, we used to use the current
time as the animation's starting time.  This made it
impossible to reliably make multiple animations proceed in
lockstep, as there's an unavoidable delay between when each
animation is scheduled.  As a result, we'd end up with ugly
effects like the small empty black gap that'd often be
visible between windows in the Alt-Tab animation.  (Note that
the black gap still exists after this change, but it's now an
artifact of the way that the current LayoutManager does
animations rather than inherent to the compositor -- the gap
is gone in my in-progress LayoutManager rewrite).

This change makes us cache a timestamp the first time that
an animation is added between frames and reuse it for all
subsequently-added animations so they'll all be
synchronized.

Change-Id: I0182a0622cbb84541035535142df0c75e05b937b

BUG=chromium-os:13963
TEST=added a test; also manual inspection that there's no empty gap visible between windows on alt-tab with my not-yet-committed LayoutManager rewrite

Review URL: http://codereview.chromium.org/6880138