wm: Handle missing UnmapNotify events.
Some crash reports and warning messages make me suspect that
we're sometimes not seeing UnmapNotify events for mapped
windows before they get destroyed.  This can lead to all
kinds of problems; in this particular case, FocusManager
appears to be dereferencing a stale pointer to a Window
object.  This change makes us try to handle this case as
gracefully as we can.

BUG=chromium-os:13792
TEST=wrote a test (but note that there are no repro steps for this; it's a pretty uncommon crash)

Change-Id: Ic11b1c9a54eab92fb680aa9c2f2b75477cdb08dc