wm: Don't mask new portion of overlay on screen resize.
This fixes a bug where the code to enable or disable
compositing was leaving part of the compositing overlay
window masked out (via the shape extension) after a screen
resize.  This change makes us instead clear the shaped
region when compositing is enabled so that we don't need to
worry about updating the region after every resize.

BUG=chromium-os:14951
TEST=manual (plugged/unplugged larger external monitor while a context menu was open and checked that the screen was resized as expected) and added a unit test

Change-Id: I24ca27933e194bc62087fdfb716171e92caa8401
Reviewed-on: http://gerrit.chromium.org/gerrit/408
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>