wm: Make window move/resize code simpler and more robust.
I had previously written tricky code that kept track of the
gravity used during a resize and tried to use that to make
an informed decision about how to update the actor's
position.  This falls down when using SetBounds() to do more
complicated simultaneous move/resizes where none of the
window's edges remains fixed.

Much simpler is to track a window's requested composited
position but hold off entirely on updating the actor's
position until the actor's size matches that of the
underlying X window.

BUG=chromium-os:13963
TEST=added; also tried it and didn't notice anything being obviously broken

Change-Id: Ibeaba33397dcac76f8ed655506e47d781c975d6f
Reviewed-on: http://gerrit.chromium.org/gerrit/1637
Reviewed-by: Greg Spencer <gspencer@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>