wm: Set root window's cursor at startup.
Chrome used to do this, but it was removed by
http://codereview.chromium.org/6973029/, which unified some
login code.  Our X server is patched to clear the root
window's cursor at startup.  As a result, windows that
didn't have a cursor set were inheriting the empty cursor
from the root.  Setting the cursor in the window manager (as
we used to do a long time ago) fixes this.

BUG=chromium-os:15767
TEST=manual: did a regular build; checked that the cursor is still hidden at startup and that it appears as soon as i move it and doesn't disappear when moved over certain windows

Change-Id: I5737f34856fbfbd7ff75e7d9ddc861f2e57280b8
Reviewed-on: http://gerrit.chromium.org/gerrit/1559
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>